requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyinstaller==6.10.0
click==8.1.7
pytest==8.3.2
python-dateutil~=2.9.0.post0